requirements.txt
alembic==1.10.4
aniso8601==9.0.1
anyio==3.6.2
black==23.3.0
click==8.1.3
fastapi==0.95.1
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
psycopg2==2.9.6
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
singledispatch==4.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
starlette-graphene3==0.6.0
strawberry-graphql==0.177.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.22.0
